Created AddMilestoneItemModal component
